# src/main/java/com/example/bankcards/dto/AuthRequest.java
package com.example.bankcards.dto;


import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

@Schema(description = "Запрос на аутентификацию")
public class AuthRequest {
    @Email(message = "Некорректный email")
    @NotBlank(message = "Email обязателен")
    @Schema(description = "Email пользователя", example = "ivanov@example.com")
    private String email;

    @NotBlank(message = "Пароль обязателен")
    @Schema(description = "Пароль пользователя", example = "password123")
    private String password;

    public String getEmail() {
        return email;
    }

    public String getPassword() {
        return password;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}