# src/main/java/com/example/bankcards/dto/ErrorResponse.java
package com.example.bankcards.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Getter;
import lombok.RequiredArgsConstructor;

@Getter
@RequiredArgsConstructor
public class ErrorResponse {

    @Schema(description = "Сообщение об ошибке", example = "Пользователь не найден")
    private final String message;
}