# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Bank REST API
  version: 1.0.0
  description: REST API для управления пользователями, картами и транзакциями
  contact:
    name: Andrey Khimchenko
    email: khimchenko_andrey@mail.ru

servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /api/auth/register:
    post:
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Пользователь с таким email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверный email или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/me:
    get:
      summary: Получение профиля текущего пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      summary: Обновление профиля текущего пользователя
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Обновленный профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /api/admin/users:
    get:
      summary: Получение списка всех пользователей (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      summary: Создание пользователя (ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Созданный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /api/admin/users/{id}:
    get:
      summary: Получение пользователя по ID (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      summary: Обновление пользователя (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Обновленный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      summary: Удаление пользователя (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Пользователь удален

  /api/admin/cards:
    get:
      summary: Получение всех карт (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: Страница карт
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer

  /api/admin/cards/user/{userId}:
    get:
      summary: Получение карт пользователя по ID (ADMIN)
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: ID пользователя
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          description: Номер страницы
        - in: query
          name: size
          schema:
            type: integer
            default: 5
          description: Размер страницы
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /api/admin/cards/{userId}:
    post:
      summary: Создание карты для пользователя (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Карта создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/admin/cards/{cardId}/block:
    put:
      summary: Блокировка карты (ADMIN)
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта заблокирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/cards/{cardId}/activate:
    put:
      summary: Активация карты (ADMIN)
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта активирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/cards/{cardId}:
    delete:
      summary: Удаление карты (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Карта удалена

  /api/users/me/cards:
    get:
      summary: Получение карт текущего пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список карт
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardResponse'

  /api/users/me/cards/{cardId}/balance:
    get:
      summary: Получение баланса карты
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Баланс карты
          content:
            application/json:
              schema:
                type: number
                format: double

  /api/users/me/cards/{cardId}/request-block:
    put:
      summary: Запрос на блокировку карты (USER)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Запрос на блокировку отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/transactions/my:
    get:
      summary: Получение транзакций текущего пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список транзакций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'

  /api/transactions/transfer:
    post:
      summary: Перевод между своими картами
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Результат перевода
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/transactions/all:
    get:
      summary: Получение всех транзакций (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Все транзакции
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'

  /api/transactions/user/{userId}:
    get:
      summary: Получение транзакций конкретного пользователя (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Транзакции пользователя
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phoneNumber: { type: string }
        password: { type: string }
        role: { type: string }
    AuthRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        message: { type: string }
        user:
          $ref: '#/components/schemas/UserResponse'
    UserRequest:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phoneNumber: { type: string }
        role: { type: string }
    UserResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        role: { type: string }
    CreateCardRequest:
      type: object
      properties:
        cardNumber: { type: string }
        balance: { type: number, format: double }
    CardResponse:
      type: object
      properties:
        id: { type: integer }
        cardNumber: { type: string }
        cardHolder: { type: string }
        expirationDate: { type: string, format: date }
        cardStatus: { type: string }
        balance: { type: number, format: double }
        blockRequested: { type: boolean }
    TransferRequest:
      type: object
      properties:
        sourceCardId: { type: integer }
        destinationCardId: { type: integer }
        amount: { type: number, format: double }
        description: { type: string }
    TransactionResponse:
      type: object
      properties:
        id: { type: integer }
        sourceCardMasked: { type: string }
        destinationCardMasked: { type: string }
        amount: { type: number, format: double }
        type: { type: string }
        status: { type: string }
        timestamp: { type: string, format: date-time }
        description: { type: string }
    ErrorResponse:
      type: object
      properties:
        message: { type: string }